requirements.txt
textual>=0.76.0
rich>=13.7.0
python-dotenv
requests